Modified bayesian.jpi file for CVE notification
2 new files were generated for direct dependencies and transitive dependencies.
The dependency files are used for CVE notification. The plugin also has an additional
logic where it would not call the scan API if the notification flag is switched off.
This commit is to affect that change and the .hpi file (renamed to .jpi) file was added
to this repo.

The .hpi file was built from fabric8-analytics/fabric8-analytics-jenkins-plugin#17

Tag: https://github.com/fabric8-analytics/fabric8-analytics-jenkins-plugin/releases/tag/0.3.0
Resolves: https://github.com/openshiftio/openshift.io/issues/3982